Seperating the Join from the Filter Predicates

+++ source/QueryOptimization/ColumnSubset.h
@@ -10,7 +10,7 @@ class ColumnSubset {
 
     List* columns;       // 0.0 -> 0.1 -> 0.2
     ColumnStatistics *interm_stats[columns->size];  // stats[0.0] -> stats[0.1] -> stats[0.2]
-    ull cost;
+    unsigned long long cost;
     
 
     public:
